requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
debugpy==1.8.14
Django==4.2.23
django-cors-headers==3.7.0
django-eventstream==4.5.1
django-extensions==3.1.5
django-filter==2.4.0
django-grip==3.5.2
django-redis==5.4.0
django-unixtimestampfield==0.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gevent==22.10.2
greenlet==2.0.2
gripcontrol==4.2.0
gunicorn==21.2.0
idna==3.10
MarkupSafe==2.1.5
numpy==1.24.3
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.8.6
pubcontrol==3.5.0
PyJWT==2.9.0
python-dotenv==1.0.0
whitenoise==6.4.0
pytz==2025.2
redis==6.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zope.event==5.0
zope.interface==7.2
psycogreen==1.0.2
django-request-logging==0.7.5
djangorestframework-api-key==3.0.0